+++ opentelemetry-sdk/tests/trace/test_sampling.py
@@ -30,10 +30,11 @@ class TestSampler(unittest.TestCase):
             ),
             0xDEADBEF1,
             0xDEADBEF2,
-            "unsampled parent, sampling on",
+            {"sampled parent":"sampling on"},
         )
-        self.assertTrue(no_record_always_on.sampled)
-        self.assertEqual(no_record_always_on.attributes, {})
+        self.assertTrue(sampling.is_sampled(no_record_always_on.decision))
+        self.assertEqual(no_record_always_on.attributes,
+            {"sampled parent":"sampling on"})
 
         sampled_always_on = sampling.ALWAYS_ON.should_sample(
             trace.SpanContext(
@@ -41,10 +42,11 @@ class TestSampler(unittest.TestCase):
             ),
             0xDEADBEF1,
             0xDEADBEF2,
-            "sampled parent, sampling on",
+            {"sampled parent":"sampling on"},
         )
-        self.assertTrue(sampled_always_on.sampled)
-        self.assertEqual(sampled_always_on.attributes, {})
+        self.assertTrue(sampling.is_sampled(no_record_always_on.decision))
+        self.assertEqual(sampled_always_on.attributes,
+            {"sampled parent":"sampling on"})
 
     def test_always_off(self):
         no_record_always_off = sampling.ALWAYS_OFF.should_sample(
@@ -55,7 +57,7 @@ class TestSampler(unittest.TestCase):
             0xDEADBEF2,
             "unsampled parent, sampling off",
         )
-        self.assertFalse(no_record_always_off.sampled)
+        self.assertFalse(sampling.is_sampled(no_record_always_off.decision))
         self.assertEqual(no_record_always_off.attributes, {})
 
         sampled_always_on = sampling.ALWAYS_OFF.should_sample(
@@ -66,7 +68,7 @@ class TestSampler(unittest.TestCase):
             0xDEADBEF2,
             "sampled parent, sampling off",
         )
-        self.assertFalse(sampled_always_on.sampled)
+        self.assertFalse(sampling.is_sampled(no_record_always_off.decision))
         self.assertEqual(sampled_always_on.attributes, {})
 
     def test_default_on(self):
@@ -78,7 +80,7 @@ class TestSampler(unittest.TestCase):
             0xDEADBEF2,
             "unsampled parent, sampling on",
         )
-        self.assertFalse(no_record_default_on.sampled)
+        self.assertFalse(sampling.is_sampled(no_record_default_on.decision))
         self.assertEqual(no_record_default_on.attributes, {})
 
         sampled_default_on = sampling.DEFAULT_ON.should_sample(
@@ -87,10 +89,19 @@ class TestSampler(unittest.TestCase):
             ),
             0xDEADBEF1,
             0xDEADBEF2,
-            "sampled parent, sampling on",
+            {"sampled parent":"sampling on"},
         )
-        self.assertTrue(sampled_default_on.sampled)
-        self.assertEqual(sampled_default_on.attributes, {})
+        self.assertTrue(sampling.is_sampled(sampled_default_on.decision))
+        self.assertEqual(sampled_default_on.attributes, {"sampled parent":"sampling on"})
+
+        default_on = sampling.DEFAULT_ON.should_sample(
+            None,
+            0xDEADBEF1,
+            0xDEADBEF2,
+            {"sampled parent":"sampling on"},
+        )
+        self.assertTrue(sampling.is_sampled(default_on.decision))
+        self.assertEqual(default_on.attributes, {"sampled parent":"sampling on"})
 
     def test_default_off(self):
         no_record_default_off = sampling.DEFAULT_OFF.should_sample(
@@ -101,7 +112,7 @@ class TestSampler(unittest.TestCase):
             0xDEADBEF2,
             "unsampled parent, sampling off",
         )
-        self.assertFalse(no_record_default_off.sampled)
+        self.assertFalse(sampling.is_sampled(no_record_default_off.decision))
         self.assertEqual(no_record_default_off.attributes, {})
 
         sampled_default_off = sampling.DEFAULT_OFF.should_sample(
@@ -110,10 +121,19 @@ class TestSampler(unittest.TestCase):
             ),
             0xDEADBEF1,
             0xDEADBEF2,
-            "sampled parent, sampling off",
+            {"sampled parent":"sampling on"},
         )
-        self.assertTrue(sampled_default_off.sampled)
-        self.assertEqual(sampled_default_off.attributes, {})
+        self.assertTrue(sampling.is_sampled(sampled_default_off.decision))
+        self.assertEqual(sampled_default_off.attributes, {"sampled parent":"sampling on"})
+
+        default_off = sampling.DEFAULT_OFF.should_sample(
+            None,
+            0xDEADBEF1,
+            0xDEADBEF2,
+            "unsampled parent, sampling off",
+        )
+        self.assertFalse(sampling.is_sampled(default_off.decision))
+        self.assertEqual(default_off.attributes, {})
 
     def test_probability_sampler(self):
         sampler = sampling.ProbabilitySampler(0.5)
@@ -121,59 +141,38 @@ class TestSampler(unittest.TestCase):
         # Check that we sample based on the trace ID if the parent context is
         # null
         self.assertTrue(
-            sampler.should_sample(
-                None, 0x7FFFFFFFFFFFFFFF, 0xDEADBEEF, "span name"
-            ).sampled
+            sampling.is_sampled(
+                sampler.should_sample(
+                    None, 0x7FFFFFFFFFFFFFFF, 0xDEADBEEF, "span name"
+                ).decision
+            )
         )
         self.assertFalse(
-            sampler.should_sample(
-                None, 0x8000000000000000, 0xDEADBEEF, "span name"
-            ).sampled
-        )
-
-        # Check that the sampling decision matches the parent context if given,
-        # and that the sampler ignores the trace ID
-        self.assertFalse(
-            sampler.should_sample(
-                trace.SpanContext(
-                    0xDEADBEF0,
-                    0xDEADBEF1,
-                    is_remote=False,
-                    trace_flags=TO_DEFAULT,
-                ),
-                0x7FFFFFFFFFFFFFFF,
-                0xDEADBEEF,
-                "span name",
-            ).sampled
-        )
-        self.assertTrue(
-            sampler.should_sample(
-                trace.SpanContext(
-                    0xDEADBEF0,
-                    0xDEADBEF1,
-                    is_remote=False,
-                    trace_flags=TO_SAMPLED,
-                ),
-                0x8000000000000000,
-                0xDEADBEEF,
-                "span name",
-            ).sampled
+            sampling.is_sampled(
+                sampler.should_sample(
+                    None, 0x8000000000000000, 0xDEADBEEF, "span name"
+                ).decision
+            )
         )
 
     def test_probability_sampler_zero(self):
         default_off = sampling.ProbabilitySampler(0.0)
         self.assertFalse(
-            default_off.should_sample(
-                None, 0x0, 0xDEADBEEF, "span name"
-            ).sampled
+            sampling.is_sampled(
+                default_off.should_sample(
+                    None, 0x0, 0xDEADBEEF, "span name"
+                ).decision
+            )
         )
 
     def test_probability_sampler_one(self):
         default_off = sampling.ProbabilitySampler(1.0)
         self.assertTrue(
-            default_off.should_sample(
-                None, 0xFFFFFFFFFFFFFFFF, 0xDEADBEEF, "span name"
-            ).sampled
+            sampling.is_sampled(
+                default_off.should_sample(
+                    None, 0xFFFFFFFFFFFFFFFF, 0xDEADBEEF, "span name"
+                ).decision
+            )
         )
 
     def test_probability_sampler_limits(self):
@@ -183,14 +182,18 @@ class TestSampler(unittest.TestCase):
         # should get sampled.
         almost_always_off = sampling.ProbabilitySampler(2 ** -64)
         self.assertTrue(
-            almost_always_off.should_sample(
-                None, 0x0, 0xDEADBEEF, "span name"
-            ).sampled
+            sampling.is_sampled(
+                almost_always_off.should_sample(
+                    None, 0x0, 0xDEADBEEF, "span name"
+                ).decision
+                )
         )
         self.assertFalse(
-            almost_always_off.should_sample(
-                None, 0x1, 0xDEADBEEF, "span name"
-            ).sampled
+            sampling.is_sampled(
+                almost_always_off.should_sample(
+                    None, 0x1, 0xDEADBEEF, "span name"
+                ).decision
+            )
         )
         self.assertEqual(
             sampling.ProbabilitySampler.get_bound_for_rate(2 ** -64), 0x1
@@ -206,9 +209,11 @@ class TestSampler(unittest.TestCase):
 
         almost_always_on = sampling.ProbabilitySampler(1 - 2 ** -64)
         self.assertTrue(
-            almost_always_on.should_sample(
-                None, 0xFFFFFFFFFFFFFFFE, 0xDEADBEEF, "span name"
-            ).sampled
+            sampling.is_sampled(
+                almost_always_on.should_sample(
+                    None, 0xFFFFFFFFFFFFFFFE, 0xDEADBEEF, "span name"
+                ).decision
+            )
         )
 
         # These tests are logically consistent, but fail because of the float
@@ -234,13 +239,51 @@ class TestSampler(unittest.TestCase):
             1 - sys.float_info.epsilon
         )
         self.assertFalse(
-            almost_almost_always_on.should_sample(
-                None, 0xFFFFFFFFFFFFFFFF, 0xDEADBEEF, "span name"
-            ).sampled
+            sampling.is_sampled(
+                almost_almost_always_on.should_sample(
+                    None, 0xFFFFFFFFFFFFFFFF, 0xDEADBEEF, "span name"
+                ).decision
+            )
         )
         # Check that the higest effective sampling rate is actually lower than
         # the highest theoretical sampling rate. If this test fails the test
         # above is wrong.
         self.assertLess(
             almost_almost_always_on.bound, 0xFFFFFFFFFFFFFFFF,
+        )
+
+    def test_parent_or_else(self):
+        sampler = sampling.ParentOrElse(sampling.ALWAYS_ON)
+        # Check that the sampling decision matches the parent context if given
+        self.assertFalse(
+            sampling.is_sampled(
+                sampler.should_sample(
+                    trace.SpanContext(
+                        0xDEADBEF0,
+                        0xDEADBEF1,
+                        is_remote=False,
+                        trace_flags=TO_DEFAULT,
+                    ),
+                    0x7FFFFFFFFFFFFFFF,
+                    0xDEADBEEF,
+                    "span name",
+                ).decision
+            )
+        )
+
+        sampler2 = sampling.ParentOrElse(sampling.ALWAYS_OFF)
+        self.assertTrue(
+            sampling.is_sampled(
+                sampler2.should_sample(
+                    trace.SpanContext(
+                        0xDEADBEF0,
+                        0xDEADBEF1,
+                        is_remote=False,
+                        trace_flags=TO_SAMPLED,
+                    ),
+                    0x8000000000000000,
+                    0xDEADBEEF,
+                    "span name",
+                ).decision
+            )
         )